fix model transaction, notification follow  many to many and polymorphism
