+++ yashoes/admin.py
@@ -3,7 +3,7 @@ from django.contrib.auth.backends import get_user_model
 from django.contrib.auth.admin import UserAdmin
 from .forms import CustomUserChangeForm, CustomUserCreationForm
 from .models import User
-from yashoes.model import user_variant, category, comment, notification, product_category, product, rating, transaction_variant, transaction, variant
+from yashoes.model import user_variant, category, comment, notification, product_category, product, rating, transaction, variant
 from django.utils.translation import gettext_lazy as _
 from django.utils.safestring import mark_safe
