+++ yashoes/model/transaction.py
@@ -1,5 +1,6 @@
 from django.db import models
 from django.conf import settings
+from yashoes.model.variant import Variant
 
 
 class Transaction(models.Model):
@@ -15,9 +16,20 @@ class Transaction(models.Model):
     status = models.CharField(max_length=1, choices=status_choices)
     total = models.BigIntegerField()
     phone_number = models.CharField(max_length=10)
+    variants = models.ManyToManyField(Variant, through='TransactionVariant')
     created_at = models.DateTimeField(auto_now_add=True, blank=True)
     updated_at = models.DateTimeField(auto_now_add=True, blank=True)
     deleted_at = models.DateTimeField(null=True, blank=True)
 
     class Meta:
         db_table = "transaction"
+
+class TransactionVariant(models.Model):
+    transaction = models.ForeignKey(Transaction, on_delete=models.CASCADE)
+    variant = models.ForeignKey(Variant, on_delete=models.CASCADE)
+    quantity = models.IntegerField(default=0)
+    created_at = models.DateTimeField(auto_now_add=True, blank=True)
+    updated_at = models.DateTimeField(auto_now_add=True, blank=True)
+    deleted_at = models.DateTimeField(null=True, blank=True)
+    class Meta:
+        db_table = "transactions_variants"
