+++ yashoes/user/views.py
@@ -29,5 +29,5 @@ class UserView(viewsets.ViewSet):
             'username': user.username,
             'address': user.address,
             'phone_number': user.phone_number,
-            'image_profile': user.image_profile.url,
+            'image_profile': user.image_profile.name if not user.image_profile.name else user.image_profile.url,
         }, status=200)
